// src/enums/Face.ts
export enum Face {
  N2 = "2",
  N3 = "3",
  N4 = "4",
  N5 = "5",
  N6 = "6",
  N7 = "7",
  N8 = "8",
  N9 = "9",
  N10 = "10",
  J = "J",
  Q = "Q",
  K = "K",
  A = "A",
}

export const FACES = [
  Face.N2,
  Face.N3,
  Face.N4,
  Face.N5,
  Face.N6,
  Face.N7,
  Face.N8,
  Face.N9,
  Face.N10,
  Face.J,
  Face.Q,
  Face.K,
  Face.A,
];